+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/Extensions.kt
@@ -11,6 +11,10 @@ import at.asitplus.wallet.healthid.HealthIdScheme
 import at.asitplus.wallet.idaustria.IdAustriaScheme
 import at.asitplus.wallet.lib.data.AttributeIndex
 import at.asitplus.wallet.lib.data.ConstantIndex
+import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation
+import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.ISO_MDOC
+import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.PLAIN_JWT
+import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.SD_JWT
 import at.asitplus.wallet.lib.data.dif.InputEvaluator
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
@@ -20,29 +24,29 @@ import at.asitplus.wallet.taxid.TaxIdScheme
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.encodeToJsonElement
 
-fun InputDescriptor.extractConsentData(): Triple<ConstantIndex.CredentialRepresentation, ConstantIndex.CredentialScheme, Map<NormalizedJsonPath, Boolean>> {
+fun InputDescriptor.extractConsentData(): Triple<CredentialRepresentation, ConstantIndex.CredentialScheme, Map<NormalizedJsonPath, Boolean>> {
     val inputDescriptor = this
-    val credentialRepresentation = if (inputDescriptor.format?.sdJwt != null || inputDescriptor.format?.jwtSd != null) {
-        ConstantIndex.CredentialRepresentation.SD_JWT
-    } else if (inputDescriptor.format?.msoMdoc != null) {
-        ConstantIndex.CredentialRepresentation.ISO_MDOC
-    } else {
-        ConstantIndex.CredentialRepresentation.PLAIN_JWT
+    val credentialRepresentation = when {
+        inputDescriptor.format?.sdJwt != null || inputDescriptor.format?.jwtSd != null -> SD_JWT
+        inputDescriptor.format?.msoMdoc != null -> ISO_MDOC
+        else -> PLAIN_JWT
     }
     val pattern = when (credentialRepresentation) {
-        ConstantIndex.CredentialRepresentation.SD_JWT ->
-            inputDescriptor.constraints?.fields?.firstOrNull() { it.path.toString().contains("vct") }?.filter?.let {
-                it.pattern ?: it.const
-            }
-        ConstantIndex.CredentialRepresentation.ISO_MDOC -> inputDescriptor.id
-        ConstantIndex.CredentialRepresentation.PLAIN_JWT -> throw Throwable("PLAIN_JWT not implemented")
+        SD_JWT -> inputDescriptor.constraints?.fields?.firstOrNull {
+            it.path.toString().contains("vct")
+        }?.filter?.let {
+            it.pattern ?: it.const
+        }
+        ISO_MDOC -> inputDescriptor.id
+        PLAIN_JWT -> throw Throwable("PLAIN_JWT not implemented")
     } ?: throw Throwable("Missing Pattern")
 
     val scheme = AttributeIndex.schemeSet.firstOrNull {
         when (credentialRepresentation) {
-            ConstantIndex.CredentialRepresentation.PLAIN_JWT -> throw Throwable("PLAIN_JWT not implemented")
-            ConstantIndex.CredentialRepresentation.SD_JWT -> it.sdJwtType == pattern
-            ConstantIndex.CredentialRepresentation.ISO_MDOC -> it.isoDocType == pattern
+            PLAIN_JWT -> throw Throwable("PLAIN_JWT not implemented")
+            // This is not entirely correct, but we'll try to work around incorrect definitions in our credentials
+            SD_JWT -> it.sdJwtType == pattern || it.isoNamespace == pattern
+            ISO_MDOC -> it.isoDocType == pattern
         }
     } ?: throw Throwable("Missing scheme")
 
@@ -53,9 +57,8 @@ fun InputDescriptor.extractConsentData(): Triple<ConstantIndex.CredentialReprese
     ).getOrNull() ?: throw Throwable("Unable to evaluate constraints")
 
     val attributes = constraintsMap.mapNotNull {
-        val path = it.value.map { value ->
-            value.normalizedJsonPath
-        }.firstOrNull() ?: return@mapNotNull null
+        val path = it.value.map { it.normalizedJsonPath }.firstOrNull()
+            ?: return@mapNotNull null
         val optional = it.key.optional ?: false
         path to optional
     }.toMap()
@@ -64,10 +67,11 @@ fun InputDescriptor.extractConsentData(): Triple<ConstantIndex.CredentialReprese
 }
 
 fun ConstantIndex.CredentialScheme.toJsonElement(
-    representation: ConstantIndex.CredentialRepresentation,
+    representation: CredentialRepresentation,
 ): JsonElement {
     val dataElements = when (this) {
         ConstantIndex.AtomicAttribute2023, IdAustriaScheme, EuPidScheme, MobileDrivingLicenceScheme, HealthIdScheme -> this.claimNames
+        // TODO Use: this.claim names for all schemes
         PowerOfRepresentationScheme -> PowerOfRepresentationDataElements.ALL_ELEMENTS
         CertificateOfResidenceScheme -> CertificateOfResidenceDataElements.ALL_ELEMENTS
         CompanyRegistrationScheme -> CompanyRegistrationDataElements.ALL_ELEMENTS
@@ -76,15 +80,10 @@ fun ConstantIndex.CredentialScheme.toJsonElement(
     }
 
     return dataElements.associateWith { "" }.let { attributes ->
-        if (representation != ConstantIndex.CredentialRepresentation.ISO_MDOC) {
-            val representationConstraint = when (representation) {
-                ConstantIndex.CredentialRepresentation.PLAIN_JWT -> "type" to this.vcType
-                ConstantIndex.CredentialRepresentation.SD_JWT -> "vct" to this.sdJwtType
-                ConstantIndex.CredentialRepresentation.ISO_MDOC -> "" to this.isoDocType //case can never happen
-            }
-            vckJsonSerializer.encodeToJsonElement(attributes + representationConstraint)
-        } else {
-            vckJsonSerializer.encodeToJsonElement(mapOf(this.isoNamespace to attributes))
+        when (representation) {
+            PLAIN_JWT -> vckJsonSerializer.encodeToJsonElement(attributes + ("type" to this.vcType))
+            SD_JWT -> vckJsonSerializer.encodeToJsonElement(attributes + ("vct" to this.sdJwtType))
+            ISO_MDOC -> vckJsonSerializer.encodeToJsonElement(mapOf(this.isoNamespace to attributes))
         }
     }
 }